Refactoring of project tree
