+++ app/src/main/java/ru/geekbrains/noteapp/viewmodel/viewmodel/CustomViewModel.kt
@@ -3,7 +3,7 @@ package ru.geekbrains.noteapp.viewmodel.viewmodel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import ru.geekbrains.noteapp.viewmodel.viewstate.CustomViewState
+import ru.geekbrains.noteapp.viewstate.CustomViewState
 
 open class CustomViewModel<T, VS : CustomViewState<T>> : ViewModel() {
