+++ app/src/main/java/ru/geekbrains/noteapp/viewstate/BaseViewState.kt
@@ -1,6 +1,5 @@
-package ru.geekbrains.noteapp.viewmodel.viewstate
+package ru.geekbrains.noteapp.viewstate
 
 import ru.geekbrains.noteapp.model.data.Note
-import ru.geekbrains.noteapp.viewmodel.viewstate.CustomViewState
 
 class BaseViewState(val notes: List<Note>? = null, error: Throwable? = null) : CustomViewState<List<Note>?>(notes, error)
