+++ app/src/main/java/ru/geekbrains/noteapp/viewstate/CustomViewState.kt
@@ -1,3 +1,3 @@
-package ru.geekbrains.noteapp.viewmodel.viewstate
+package ru.geekbrains.noteapp.viewstate
 
 open class CustomViewState<T>(val data: T?, val error: Throwable?)
